requirements.txt
Django==2.2.6
pytz==2020.1
sqlparse==0.3.1
Django==2.2.6
django-crispy-forms==1.9.2
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
